feat(game): updating map connector for levels

+++ game/TwitchRealms/Assets/Scripts/BubbleData.cs
@@ -8,6 +8,6 @@ public class BubbleData
     public float x = 0f;
     public float y = 0f;
     public string bubbleColor = "#00FF00";
-    public float bubbleSize = 50;
+    public int bubbleSize = 50;
     public string userId = null;
 }
